% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KStest.R
\name{makeModelMatrix}
\alias{makeModelMatrix}
\title{makeModelMatrix: Create over-parameterized model matrix for tsBART fitting.}
\usage{
makeModelMatrix(df)
}
\arguments{
\item{df}{A data frame. Defaults to NULL.}
}
\value{
\item{X}{An overparameterized model matrix, where factor covariates are split into columns of indicators, with one column for each level, including the baseline level.}
}
\description{
This function creates a model matrix for use with tsbart().
}
\examples{
\dontrun{
## Create model matrix.
x <- makeModelMatrix(mydf)

}}
\author{
Jennifer E. Starling
}
\keyword{tsbart}


