% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtnorm}
\alias{rtnorm}
\title{Pseudorandom numbers from a Gaussian distribution that is truncated to an interval.}
\usage{
rtnorm(lower, upper, mu, sigma)
}
\arguments{
\item{lower}{lower bound of truncation interval}

\item{upper}{upper bound of truncation interval}

\item{mu}{mean of the normal random variable (before truncation)}

\item{sigma}{standard deviation of the normal random variable (before truncation)}
}
\description{
Pseudorandom numbers from a Gaussian distribution that is truncated to an interval.
}
